restructure graphics lib and system

+++ Source/Math/Vector.cpp
@@ -144,20 +144,20 @@ Math::Vec3::Vec3(const Vec4 & v) : x(v.x), y(v.y), z(v.z)
 Math::Vec3::Vec3(const Quat & q)
 {
     // roll
-    float sinr_cosp = +2.0 * (q.w * q.x + q.y * q.z);
-    float cosr_cosp = +1.0 - 2.0 * (q.x * q.x + q.y * q.y);
+    float sinr_cosp = +2.0f * (q.w * q.x + q.y * q.z);
+    float cosr_cosp = +1.0f - 2.0f * (q.x * q.x + q.y * q.y);
     y = atan2(sinr_cosp, cosr_cosp);
 
     // pitch
-    double sinp = +2.0 * (q.w * q.y - q.z * q.x);
+    float sinp = +2.0f * (q.w * q.y - q.z * q.x);
     if (fabs(sinp) >= 1)
-        x = copysign(Math::PI / 2, sinp); // use 90 degrees if out of range
+        x = copysign(Math::PI / 2.0f, sinp); // use 90 degrees if out of range
     else
         x = asin(sinp);
 
     // yaw 
-    double siny_cosp = +2.0 * (q.w * q.z + q.x * q.y);
-    double cosy_cosp = +1.0 - 2.0 * (q.y * q.y + q.z * q.z);
+    double siny_cosp = +2.0f * (q.w * q.z + q.x * q.y);
+    double cosy_cosp = +1.0f - 2.0f * (q.y * q.y + q.z * q.z);
     z = atan2(siny_cosp, cosy_cosp);
 }
